refactor: 💡 localStorage.ts renamed to storage.ts
